+++ core/templates/tree.ts
@@ -13,25 +13,31 @@ export default function createTree(files: EntryMetadata[]): Tree {
 
   for (const file of files) {
     const { label, name } = file;
-    placeFile(tree, name, label.split("/"));
+    placeFile(tree, name, label.split("/"), name.split("/"));
   }
 
   return tree;
 }
 
-function placeFile(tree: Tree, name: string, parts: string[]) {
-  const part = parts.shift()!;
+function placeFile(
+  tree: Tree,
+  name: string,
+  labelParts: string[],
+  nameParts: string[],
+) {
+  const labelPart = labelParts.shift()!;
 
-  if (parts.length) {
+  if (labelParts.length) {
+    const namePart = nameParts.shift()!;
     tree.folders ??= new Map();
-    const folder: Tree = tree.folders.get(part) ?? {
-      path: `${tree.path}${part}/`,
+    const folder: Tree = tree.folders.get(labelPart) ?? {
+      path: `${tree.path}${namePart}/`,
     };
-    placeFile(folder, name, parts);
-    tree.folders.set(part, folder);
+    placeFile(folder, name, labelParts, nameParts);
+    tree.folders.set(labelPart, folder);
     return;
   }
 
   tree.files ??= new Map();
-  tree.files.set(part, name);
+  tree.files.set(labelPart, name);
 }